requirements.txt
diff-match-patch==20230430
edge-tts>=2.2.2,<8.0.0
numpy==1.22.0
pandas==1.5.3
pydub==0.25.1
python-dotenv==0.21.1
streamlit==1.21.0
torch==2.1.0
transformers==4.33.3
TTS==0.22.0
pydub
